+++ tests/bootstrap.php
@@ -15,6 +15,3 @@ use Tester\Environment;
 
 // Nette Tester initialization.
 Environment::setup();
-
-// This is done in this version only. See FirebaseJwtDecoder for info.
-error_reporting(E_ALL ^ E_USER_WARNING);